\name{Beta}
\alias{Beta}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Extended Beta Family}
\description{
  A function used in the \code{hglm} package which extends the usage of the Beta family.
}
\usage{
Beta(link = "logit")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{link}{the link function}
}

\value{
  Output as for other GLM families 
}
