\name{cancer}
\docType{data}
\alias{cancer}
\alias{O}
\alias{E}
\alias{Paff}
\alias{nbr}
\title{Scottish lip cancer dataset from Clayton and Kaldor (1987)}
\description{
The Scottish lip cancer dataset. 
}
\format{ The \sQuote{cancer} dataset contains 4 objects as follows. 
 \describe{
 \item{O}{Observed frequency.}
 \item{E}{Offset.}
 \item{Paff}{Fixed effects.}
 \item{nbr}{Spatial correlation matrix D for CAR model.}
}
}
\source{Clayton D, Kaldor J 1987. Empirical Bayes Estimation of Age-standardized Relative Risk for use in Disease Mapping. \emph{Biometrics} 43, 671--681}
\references{Clayton D, Kaldor J 1987. Empirical Bayes Estimation of Age-standardized Relative Risk for use in Disease Mapping. \emph{Biometrics} 43, 671--681}
\keyword{datasets}