\name{inverse.gamma}
\alias{inverse.gamma}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Inverse Gamma Family}
\description{
  A function used in the \code{hglm} package for the inverse Gamma family.
}
\usage{
inverse.gamma(link="inverse")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{link}{Link function.}
}

\value{
  Output as for other GLM families 
}
