\name{inverse.sqrt}
\alias{inverse.sqrt}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Inverse Square Root Family}
\description{
  A function used in the \code{hglm} package for the inverse square root family.
}
\usage{
inverse.sqrt()
}
%- maybe also 'usage' for other objects documented here.

\value{
  Output as for other GLM families 
}
