\name{print.hglm}
\alias{print.hglm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ A print method for hglm objects. }
\description{
  It prints some basic statistics from \code{hglm} estimation in a simplified way.  
}
\usage{
print.hglm(Object, ...)
}
\seealso{\link{hglm}, \link{hglm.default}, \link{hglm.formula}}

