\name{pump}
\docType{data}
\alias{pump}
\title{Pump reliability data set from Gaver and O'Muircheartaigh (1987)}
\description{
The \sQuote{pump} data set presents the failures of pumps in several systems of the water reactor neuclear plant Farley 1. 
}
\format{ The \sQuote{pump} data set contans 4 columns and 10 rows. A short description of the data columns are given below. 
 \describe{
 \item{System}{The system number.}
 \item{S}{Number of pumps failures.}
 \item{t}{Time (in thousand hours) of operation.}
 \item{Gr}{Pump groups; two levels: 1 = operated continuously, 0 = operated intermittently.}
}
}
\source{Gaver, D P. and O'Muircheartaigh, I. G. 1987. Robust Empirical Bayes Analyses of Event Rates, \emph{Technometrics} 29(1),1--15 }
\references{Lee, Y. and Nelder, J. A. 1996. Hierarchical generalized linear models, \emph{Journal of the Royal Statistical Association (B, Theory and Methods)} 58(4), 619--678.}
\keyword{datasets}