\name{salamander}
\docType{data}
\alias{salamander}
\title{Salamander mating data set from McCullagh and Nelder (1989)}
\description{
\sQuote{salamader} data set presents the outcome of an experiment which was conduceted at the University of Chicago in 1986 to study the extent at which mountain dusky salamanders from different populations would interbred. More detailed description of the data is given in its original source, McCullagh and Nelder (1989).  
}
\format{\sQuote{salamander} data set contains 6 columns and 360 rows. A brief description of the data columns is given below.
\describe{
\item{Season}{The seasons, Spring and Summer of 1986, when the experiment was carried out.}
\item{Experiment}{Experiment number; 1,2,3.}
\item{TypeM}{Type of the male salamander; Rough Butt=R and White Side=W}
\item{TypeF}{Type of the female salamander; Rough Butt=R and White Side=W}
\item{Cross}{Cross between female and male type e.g. Cross=WR mean a White Side female was crossed with a Rough Butt male.}
\item{Male}{Identification number of the male salamander.}
\item{Female}{Identification number of the female salamander.}
\item{Mate}{Whether a mating was observed, Yes=1 and No=0.}
}
}
\source{McCullagh P. and Nelder, J. A. 1989. \emph{Generalized Linear Models}, Section 14.5, Chapman and Hall/CRC.}
\keyword{datasets}