\name{inverse.gamma}
\alias{inverse.gamma}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Extended Inverse Gamma Family}
\description{
  A function used in the \code{hglm} package which extends the usage of the Inverse Gamma family.
}
\usage{
inverse.gamma(link = "inverse")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{link}{the link function}
}

\value{
  Output as for other GLM families 
}