\name{inverse.sqrt}
\alias{inverse.sqrt}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Inverse Square-root Family}
\description{
  A function used in the \code{hglm} package which creates the inverse sqrt family.
}
\usage{
inverse.sqrt()
}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%  \item{link}{the link function}
%}

\value{
  Output as for other GLM families 
}
