\name{print.hglm}
\alias{print.hglm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ A print method for hglm objects. }
\description{
  It prints some basic statistics from \code{\link{hglm}} estimation in a simplified way.  
}
\usage{
\method{print}{hglm}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A hglm object. }
  \item{\dots}{ Unused urguments }
}
\seealso{\code{\link{hglm}}, \code{\link{hglm.default}}, \code{\link{hglm.formula}}}

