\name{print.hglm}
\alias{print.hglm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ A print method for hglm objects. }
\description{
  It prints some basic statistics from \code{\link{hglm}} estimation in a simplified way.  
}
\usage{
print(Object, ...)
}
\seealso{\code{\link{hglm}}, \code{\link{hglm.default}}, \code{\link{hglm.formula}}}

