\name{print.summary.hglm}
\alias{print.summary.hglm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print mehtod for hglm summary }
\description{
  Print method for a \code{summary.hglm} object.
}
\usage{
summary(Object, digits = 4, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Object}{ A hglm object. }
  \item{digits}{ Number of significant digits to be printed. }
  \item{\dots}{Unused arguments}
}
